added sticks of butter as a double and new tests

+++ Baker-Biz/Refactor/Ingredient.cs
@@ -5,12 +5,12 @@ namespace BakerBiz
     public class Ingredient
     {
         public IngredientType Type { get; private set; }
-        public int AmountPerPie { get; private set; }
+        public double AmountPerPie { get; private set; }
         public int Amount { get; set; }
         public Units Units { get; private set; }
         public bool Required { get; private set; }
 
-        public Ingredient(IngredientType type, int perPieAmount, Units units, bool required = true)
+        public Ingredient(IngredientType type, double perPieAmount, Units units, bool required = true)
         {
             Type = type;
             AmountPerPie = perPieAmount;
@@ -35,9 +35,9 @@ namespace BakerBiz
             return numberOfPies;
         }
 
-        public int CalculateLeftovers(int numberOfPies)
+        public double CalculateLeftovers(int numberOfPies)
         {
-            int leftOvers = Amount - (numberOfPies * AmountPerPie);
+            double leftOvers = Amount - (numberOfPies * AmountPerPie);
 
             if (leftOvers < 0)
                 leftOvers = 0;
